pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[project]
name = "clonus"
dynamic = ["version"]
description = "partial github cloning!"
authors = [
    {name = "hex benjamin", email = "hex@hexbenjam.in"},
    {name = "Habib Rehman", email = "hi@habibrehman.com"},
]
dependencies = [
    "click>=8.1.7",
    "rich>=13.6.0",
    "python-dotenv>=1.0.0",
    "urllib3>=2.0.7",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "Apache-2.0"}

[tool.pdm.version]
source = "file"
path = "src/clonus/__init__.py"
[tool.pdm.dev-dependencies]
dev = [
    "sourcery>=1.12.0",
]
